+++ delta-coverage-core/src/main/kotlin/io/github/surpsg/deltacoverage/report/ConsoleHtmlReportLinkRenderer.kt
@@ -1,0 +1,10 @@
+package io.github.surpsg.deltacoverage.report
+
+import java.io.File
+
+object ConsoleHtmlReportLinkRenderer {
+
+    fun render(reportBound: ReportBound, reportPath: File) {
+        println("[$reportBound]: file://${reportPath.absolutePath}/index.html")
+    }
+}