+++ delta-coverage-core/src/main/kotlin/io/github/surpsg/deltacoverage/report/intellij/report/XmlReportBuilder.kt
@@ -2,17 +2,17 @@ package io.github.surpsg.deltacoverage.report.intellij.report
 
 import com.intellij.rt.coverage.report.Reporter
 import io.github.surpsg.deltacoverage.config.ReportsConfig
+import io.github.surpsg.deltacoverage.report.ReportBound
 import java.io.File
 
 internal class XmlReportBuilder(
-    reportsConfig: ReportsConfig,
-    reportBound: ReportBound,
-    reporter: Reporter,
-) : ReportBuilder(reporter, reportBound, reportsConfig) {
+    val reportBound: ReportBound,
+    private val reportsConfig: ReportsConfig,
+    private val reporter: Reporter,
+) : ReportBuilder {
 
     override fun buildReport() {
         val reportPath: File = ReportPathStrategy.Xml(reportsConfig).buildReportPath(reportBound)
         reporter.createXMLReport(reportPath)
     }
 }
-