+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # Delta-Coverage Gradle plugin Changelog
 
+## <NEXT-RELEASE>
+
+### Added
+- Render clickable HTML report path in console output.
+- Added console report. 
+
+### Fixed
+- File changes detecting when there are few classes with similar names in different packages. 
+
+
 ## 2.0.2
 
 ### Fixed